requirements.txt
discord.py==2.4.*
openai==1.57.4
PyYAML==6.0.2
python-dotenv==1.0.1
requests==2.32.3